requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.3
Flask-RESTful==0.3.9
PyMySQL==1.1.0
Werkzeug==2.3.7
Flask-Migrate==4.0.4
celery==5.3.6
redis==5.0.1
python-dotenv==1.0.0
pytz==2024.1
requests==2.31.0
Flask-JWT-Extended
cryptography
gunicorn